fix: make RRR capacity guards portable to 32-bit platforms
Comparing an int directly against math.MaxUint32 does not compile where
int is 32 bits, since the untyped constant does not fit. Widening the
accumulators to uint64 keeps the guard meaningful everywhere; both are
sums of non-negative counts, so the conversion cannot wrap.

Verified with GOARCH=386 and GOARCH=arm cross-builds.

+++ rrr.go
@@ -75,10 +75,10 @@ func NewRRR[T any](input []T, predicate func(T) bool) *RRR {
 		packClass(classes, blockIdx, class)
 	}
 
-	if totalOnes > math.MaxUint32 {
+	if uint64(totalOnes) > math.MaxUint32 {
 		panic("succincter: RRR does not support more than 2^32 ones; use Succincter for very large dense datasets")
 	}
-	if totalOffsetBits > math.MaxUint32 {
+	if uint64(totalOffsetBits) > math.MaxUint32 {
 		panic("succincter: RRR offset storage exceeds uint32 capacity")
 	}
